Add FiringCost and CanFire methods to EnergyHandler
These methods construct an EnergyLevels from a weapon and check if a ship can fire the weapon.

Since the EnergyLevels are made on the fly, this isn't a performance improvement, just an organizational improvement.

Also fixed a bug where a weapon could push a ship's shields into the negative given a negative firing cost.