Ka'sei fail and Dusk Shouting
Because the Kasei are labeled as 'NPC save', kiling them fails the mission; in which case the player doesn't get to see the `on complete` text.

This adds on `on fail` conversation that is simply the fail part.

He's emotional, so it makes sense he reverts to his most instinctive form of communication.