fix(internal): ensure UUIDs occupy initialized memory
A trivial default constructor generally doesn't initialize the memory associated with the object.
This was OK for a number of usages (including the original test code), but generally fails with our `Instantiate` pattern.
The result was that the default-constructed internal UUID value would not be considered null, and no true UUID would be generated.