Point build targets to new engine target
The game target and unit tests target now depend on the ESEngine framework.  In order to facilitate this, the project settings have been changed to place Xcode's Derived Data under build/Xcode.

SDL is no longer packaged explicitly, but instead is packaged within ESEngine (which itself is packaged explicitly).

The main build targets should take the appropriate ESEngine product depending on if they're building for Release or for Debug. (could not be tested due to warnings in ESEngine Debug becoming errors in ESEngine Release)

Changed bundle version and bundle version string to point to Current Project Version and Marketing Version, respectively.  In order to facilitate this, the Current Project Version has been set to 0.9.14 and Marketing Version points to Current Project Version.