feat(content): Break the Deep: Remnant missions out of the Bactrian line
These missions have been expanded so much over the years that they fit far better as their own line of missions instead of being tied in with the Bactrian license.

Will come back to these later to do a more thorough rework, such as filling the hole in the Bactrian line that pulling out the Remnant missions left behind, and making the Bactrian line a bit more cohesive.

Removed the mention of Rayna from the last Scientist Rescue missions since she only shows up in TMBR and Deep: Remnant, which the player may not have done. Will need to rework her character so that she's not a literal who. Really, all of these characters could do with some work.