Bugfix: Fighters should find new parents
Bug description
---------------

Fighters hang if the flagship has fighter bays available.  The fighters will race to the flagship and join.  If the flagship fills up with bays the fighters do not try to find a new parent.  Even if escorts have free bays the fighter will not try to find a new bay to join.

Expected behavior
-----------------

After a ship's bays are filled the fighter should try to find a new parent among player escorts.  If an escort has an available bay then the fighter (or drone) should board that escort.

If there are no bays among escorts then the abandoned fighter should fall back to keeping station with the player flagship.

Actual behavior
---------------

Fighters race to flagship.  When bays are filled the fighters keep station with the flagship.  Even if bays are available among escorts the fighter or drone will not attempt to board amongst a new parent.

Save game
---------

You can test with this [save game shared in discord][discord].

Testing
-------

Tested the following scenarios where a player fleet attempted to pick up two (2) abandoned fighters with various fleet configurations with free bays.

| Abandoned fighters | Flagship Free Bays | Escort Free Bays |
| ------------------ | ------------------ | ---------------- |
|                  2 |                  1 |                6 |
|                  2 |                  1 |                0 |
|                  2 |                  1 |                1 |
|                  2 |                  0 |                2 |
|                  2 |                  0 |                1 |
|                  2 |                  0 |                0 |

- In all above tested scenarios, if the flagship has a bay the fighters would attempt to board.
- If the flagship was full or the flagship has no bays then the fighters would board an escort.
- If no bays are available then the abandoned fighters would keep station with the player flagship.

Performance Impact
------------------

None.  The way `Ship::CanCarry` has been rewritten it should theoretically be more performant than the old method as the number of player escorts grows.

[discord]: https://discordapp.com/channels/251118043411775489/266345072554016768/979048779724517389