feat(ci): Update CI workflow to handle new Xcode structure
* Expand build_macos job to build all products
Add more artifacts to account for there now being multiple build targets.
Only fetch build dependencies if at least one of the products could not be cached.
Build and upload any artifacts that could not be downloaded and cached.

* Re-syntax conditional logic for build_macos job
Evaluate bool-type cache-hit result as a bool and not as compared to true

* Adopt plain variable name convention within macOS-type jobs
If a variable doesn't directly reference a binary or framework (as it usually doesn't in the macos jobs) its name shouldn't imply that it does.  SDL2_FRAMEWORK is a special case in that it doubles as a cache destination and a reference the framework’s normal location.

* Avoid quotes for xcodebuild -configuration
Also considered not using quotes for various paths, but the game target/product name contains spaces and therefore requires quotes.

* Add test_macos-unit job
Now that the Xcode project has been configured to build the unit tests, the CI should run them.

* Update the macos/Xcode path filters used by the changed job
The workspace probably shouldn't change but tests should run if it does.  Shared data needs to be checked, at least on the project level (this checks both project and workspace) in order to verify schemes.