Remove GenericName key from Linux desktop file
This key is intended to be used for cases like a music player with a
branded name where you want the user to be able to tell what it is (e.g.
"Music Player"). This doesn't really apply to games; the Comment field
which shows a longer description is more appropriate. On Linux, launcher
menu implementations often prefer the GenericName key over the Comment
when both are present. So let's remove the GenericName because the
Comment text is better and more appropriate.