Reword the explanation of that train's speed of 11D.E8, coefficients and the speed of light. I understand scientific notation but got lost relating the values here, which mixed 11D.E8, 11'DE7 and 1.1DE7.
First and foremost, the stated value 11D.E8 doesn't actually match the 1.1DE7 that we're supposedly familiar with. So fix that.

The entire talk of coefficients is really superfluous as you can just simply recognize that "11D.E7" has the same hex-digits as "11'DE7'84A". That could be all this dialogue is about. So if we're going to introduce scientific notation, we need to state "coefficient is 1.1DE7" earlier to make that idea more solid. And since we are talking coefficients, keep it simple: we don't need to see the 11'DE7'84A case at all.

It doesn't help that "E8" also means "x10^8" when the numbers we're talking about are also "x10^8". So let's make it 11D.E78, because you probably won't think that is supposed to mean "11D,000,000,000..." with 78 zeroes.

The explanation also didn't wrap everything together at the end. So I do that.

Much smaller edits:

The responses
`	"That number seems familiar."`
`	"That sounds like the speed of light."`
are silly to read, as you immediately see that the familiarity is with the speed of light. So since we're adding a digit of precision, let's address that instead with `	"Why is that number so precise?"`

Hexadecimal wouldn't have 16^7, it would be 10^7 meaning sixteen^seven.

Remove 'c' and just say 'speed of light'. No need to say 'c=X' when the text explains that's 'X' is the speed of light.

Also I'm not sure why the thousands separator is ' here , nowhere in the game do I see that. So now the numbers can be single-quoted.