Fix a subtle bug : opening the mission pane and hitting down arrow to select the first mission NOW zooms the screen to that system.
The result of 'SelectAnyMission' was being used in the wrong if statement. When it did its job and selected the first mission, it would skip over the code to iterate to the next mission (as the down key usually would). Sure. But that means it would then pass the keypress to the superclass MapPanel, and also skips over the following code that scrolled and centered on that system.

So the call to SelectAnyMission just needed to be nested deeper so that any keydown press is entirely handled in MissionPanel. (Code change does mean it's not passing the down-arrow keypress to MapPanel, which is fine since MapPanel doesn't use that key anyway)