Zoom to auto-selected mission when you change to mission panel from another panel.
I figure if this this zooming is not wanted, the mission panel probably shouldn't have auto-selected a mission at all. But while it does selection a mission for you, the screen ought to go to it - at least scroll the job list to it.

This also refactors the repeated code of selectedSystem=_,doScroll,CenterOnSystem to new method SetSelectedScrollAndCenter().