Re-work DistanceMap to be consistent: only building paths from the center outwards.
Okay this is all confusing to talk about so let's get this straight - DistanceMap was used in two ways: One way (such as when called from the Map Panel) would start the algorithm from the ship, looking for branching paths outward to many possible destinations. Another way (when a ship needed to get to a known destination) would work in reverse - starting the algorithm at the DESTINATION - finding all paths outward from the destination, until it found the path to the ship.

This two-way system simply doesn't work anymore. It assumed paths were valid to travel both ways. This is NOT VALID anymore due to the "jump range" attribute of a system. You can jump from A to B, but you might not be able to jump from B to A. So now the path edges are not bidirectional, so the reverse-path method is NOT accurate. So the DistanceMap code NEEDS to search in the direction of travel. (Confused why it wasn't like that to begin with anyway!)

So I'm overhauling DistanceMap to only work one way - starting from the start, and moving outward, matching the direction of travel.

Codewise, this amounts to changing the ship->target algorithm to simply start from the ship, instead of creating a reverse-path from the target, so now it works like it does for other uses of DistanceMap. This simplifies some parts, but has some ramifications outside, where that DistanceMap is used.

I changed variable names and comments to clarify how DistanceMap works. So now that DistanceMap is unified and single-directional, much confusion is removed - e.g. "Find paths to the given system" when it really finds paths going away from the system. "where you will go next" is NOT what the Edge was storing (for MapPanel usage at least) - it stored where the route CAME FROM. The Edges fuel/days values were for the next system, but 'danger' was a step behind. It was confusing AF to read this code! Not so much anymore.

Reworked the DistanceMap public methods to remove unneeded methods, and provided methods for code that was often used.

Added a 'destination' option for player->system constructor, which should optimize the case of shift-clicking the map to add onto an existing route.

Before, the reverse-path usage meant some code had easy access to the "first" jump that was needed. It now needs to iterate through the route back to find the first jump. This works fine though, since the entire route was needed anyway and already computed.

ShouldRefuel unified (which used DistanceMap haphazardly). The "wrapper" ShouldRefuel didn't use DistanceMap, but was doing all the same checks as the "inner" ShouldRefuel, that also used DistanceMap. But the wrapper ShouldRefuel wasn't given the entire route, only the next system, so it didn't really benefit from the inner anyway - it wasn't getting any use out of a DistanceMap to a system one jump away. So there's no reason for the wrapper. Now there's only one ShouldRefuel, and it gets the full DistanceMap to work with instead of just the next system.

Bugs that were fixed:
* Escorts wouldn't decide to jump to a system with lower "jump range" (e.g. Edusa to PostVerta with SSil Vida online)
* Escorts would think they could jump out of a system with lower "jump range", but couldn't, and would constantly land.
* Escorts wouldn't refuel when they can't reach the destination, in some cases - (Escorts were only considering the next single jump instead of the full path, when following the flagship jumping out of a system)

Some closing thoughts:

The DistanceMap fields 'route' and 'edgesTodo' store basically the same thing - known best edges. It could be unified in one structure, if the edge held both 'prev' and 'next'. Of course 'next' is only be used when building the map, at which point 'prev' isn't needed and would just be assigned to 'next' anyway. But it would make for clearer code. A big chunk of added comments explains this instead, when the structure could've just been changed. Maybe also, some code using DistanceMap would benefit from both directions being kept.

Some unoptimal edges could be processed unnecessarily as it is. When a better path is found, the Edge in route gets replaced, but the old Edge remains in edgesTodo. It'll be popped at some point, after the better edge, and will calculate edges that are guaranteed to be worse.

I suspect that escorts ordered to jump to a system don't use your known map, and can travel with whatever means necessary to get there. Not a bug I guess? But weird.

In the case for ship->target, DistanceMap is a bit excessive to work with. The DistanceMap is needed, and does stop when the destination is reached, but only the one route from ship->target is needed as a result. The rest of the DistanceMap is held around, unused. So there could be a BestRoute class that uses DistanceMap and just extracts the best route to use.