When the code determines a shop purchase fails, include the error message as part of the failure, instead of using another method to generate the error message later. This disconnect meant some failure cases didn't have an error message (and when that error message was added, it might accidentally show up for a different case (e.g. #6918)).
So now the code that determines the failure also determines the message, so there's no room for confusion.

Codewise it means the content of FailBuy is now included in CanBuy. So no more FailBuy is needed - just push the Dialog instead.

TODO: Shipyard Panel.

TODO: The same for FailSell.

When the code checks CanBuy without showing the error, perhaps now it is slower to construct an unused string. That's once per frame for the button. Anywhere CanBuy is in a loop, it breaks out on the first failure so it won't be making tons of unused strings.