Skip negative values in normal outfit attribute display.
Instead of blindly skipping attributes that are usually negative, they'll now be included when positive, with the tag "added" : e.g. "outfit space added: 15". Since they are exceptions, they'll be displayed first instead of alphabetical.

Negative values are not shown here, e.g. "missile capacity: -1" - that is already in the requirements section as "missile capacity needed: 1"