Bugfix: Smooth music playback
I am developing the plugin "Sounds of Endless Sky" which includes music by composer @Anthrophantasmus.

Bug behavior
------------

Music gets reset even if the same track is playing.  The track restarts from the beginning every time it resets.

- Jumping systems resets music.
- Pressing ESC resets music.
- Landing on planets resets music.

Desired behavior
----------------

If the track is already playing it should continue to play.  Do not reset and play from the beginning because this can be jarring.