Fixing a confusing block message
Sort of a problematic situation: The reserved cargo was appropriate for (and expects) that the player is switching to the Peregrine that they are given during this mission, so the amount of cargo their current ship has is actually irrelevant. But there's no way to tell the mission that it needs to reserve that cargo space at the *end* of the "on accept" instead of at the start of the "on offer."

So I'm simply removing it and just adding to the mission description that they need to pick up 58 tons of material. Not quite as obvious as reserving the space, but hopefully people actually read it.

Thanks to Rocky196884 on Discord, among others, for reporting this bug.