Change a news message to make sense with the Intro
Aulori is the Heliarch Consul. Elirom is the Lunarium member.
Originally was gonna have Elirom be Aulori's family name, but decided to grab it as a proper name for the Lunarium Saryd leader instead.
Forgot to check the news beforehand, apparently, so changing this so that it refers to Aulori directly.