Force full screen for Steam Deck hardware
Summary
-------

I know it isn't typical for specific hardware to be a part of the game but considering endless sky is published on steam I think it is worth directly supporting steam deck hardware.

On steam deck, the game starts maximized with a vertical pixel buffer instead of fullscreen.  This creates a blank area on the screen and causes controls to lose focus if the area is accidentally clicked.  I've been playing this regularly and miss-clicking the blank area happens very often during game play.

Feature Details
---------------

Use standard kernel APIs to read the hardware vendor and product name in order to determine if running on steam deck.  If steam deck is found, then force full screen ignoring user preference.

Testing Done
------------

* I initially tried using `Files::Read` but the way Linux kernel exposes `/sys` it is impossible to successfully read using default file functions within endless sky.
* I created a build and successfully launched this on steam deck.
* I launched on Linux (not steam deck) and user preferences for fullscreen and maximized are preserved.
* I edited `preferences.txt` and set `fullscreen 0` and `maximized 0`.  Launching the game successfully forces fullscreen with the preference `fullscreen 0`.

Performance Impact
------------------

None