variants: add No Bays variants
the only two ship I've encountered so far that has any gain from
removing the bays are the Aerie and the Bastion classes which both have internal
bays.
these bays are designed for fighter and the largest fighter (mass-wise)
in the game has the mass of 80, I deduce based on the Iron mineral that the
mass to cargo ratio is 1-1. so following the num of bays removed * heaviest
fighter in the game, I deduce the gain of the cargo the ship will receive.

Signed-off-by: Dagg Stompler <daggs@gmx.com>