map: add lionheart custom shipyard
because these variants are special and unique, much like highly custom cars
which can be found only on selected dealerships, it is logical to apply
the same logic to the ship.
so to make it really rare, the best place for this shipyard is at the
Lionheart industries main homeworld.

Signed-off-by: Dagg Stompler <daggs@gmx.com>