Update outfitter warnings for fighter
The low battery warning was changed to a low battery error.  This is because fighters will refuse to launch if less than 10 seconds of sustained flight.

There's a new low battery warning.  If there's less than 15 seconds of sustained flight, then the player is warned about low battery capacity and that battery powered flight will be a verylimited duration.