Apply suggestions from code review by Amazinite. Thanks!
I really dislike the use of fails as I find it disruptive and harder to follow. I continue to wish we could forcibly "complete" other missions in the same manner as we currently "fail" them.

I'm merging all your suggestions, then comparing the end result to a copy saved elsewhere so I can fine-toothed exam some of the content changes. most of it will stay, but I suspect there's a few bits that won't.

Co-authored-by: Amazinite <jsteck2000@gmail.com>