Remove redundant calls to GetPlanet
We already have access to `this`, and we're already grabbing the player's system directly from its variable, so might as well do the same with the planet.

I'll note that PlayerInfo is inconsistent on this, though. Sometimes functions will call GetPlanet or GetSystem, while other times they just use the planet or system variables directly.