fix(content): qualify Tomek's alleged tactical genius
Tomek is described as a tactical genius right after he's let his
emotions get the better of him and lost a major battle as a result.
These are not the actions of a tactical genius, so the description
seems odd, to say the least.

To reduce this jarring contrast, let's have Freya qualify the statement
a bit.