fix(content) razer -> razor
Razer is a gaming hardware company, not a thing you shave with or a
description of sharpness. Given that the context is a crazy person's
weird poetry, it's possible this was an intentional typo, but it still
stands out to me and seems like an error even if it was intentional,
and the correction is correct too, so let's assume it was an
unintentional typo and correct it.