Grey's review, Expedition chain part 1
At a later date will change the end portion so as to completely remove the awkward nonsense bit of acquiring outfits for them to set up base. Will instead have the player notified after a timer that the Punisher's gone back to the Coalition, and another conversation with Aulori on complete after returning to the Ring of Wisdom.