Increased heat, power draw to match x2.4 pattern
I'm hesitant to change the weapon space too much further, at least for the twin mod blaster, because the 14 mass of the twin fits together well with the 6 mass of the single, and many human ships have weapon capacities that are multiples of 20. Instead, I've increased the heat and power draw of both weapons to mass the 2.4x pattern used for the mass, and updated the twin reg's description accordingly. 
I think it also makes a bit more sense that pirates will be building stripped-down versions that punch a little bit above their weight class, but aren't as efficient as the stock models.