alternate maximum offset method
plus comments and cleaning stuff up!

so, if you were in something like a bulk freighter with a very low top speed, it would zoom forward or backward on the screen as soon as it thrusted, because it immediately reaches its top speed. Which would definitely not look right. This method instead focuses on only the problem: way too much offset pushing the ship off-screen. Instead of allowing velocity to linearly translate to visual offset, it nonlinearly translates to visual offset which tapers off as it approaches the bounds of the player's screen.