Changes to syntax in the Hai Revealed story
I'm new to github, and I hope I'm doing it all correctly!

I'm changing some of the dialogue in the Hai Revealed story because there is missing or bad syntax and some wordy or unnecessary repetition in the writing. I've also got questions and suggestions for the story but I don't know enough about Endless Sky's story style to make any changes. I'll give reasoning below. 

I've only played and read up to Hai Middle and if there is anything that will be explained that I've changed, it would be good to know. I know there is going to be a lot expanded upon but there are some things that are inconsistent to the past events in the Free Worlds. 

Changes
- I've replaced Senator with Councillor because that is how representatives in Parliament were titled before. Senator implies that Kabiru is in the Free Worlds Senate. I would also split the dialogue there and in the next few events ([Watch the clip.] or [Continue watching.]) and give the player some decisions. There is a whole lot of telling what the player should feel, instead of giving the details to make the player feel something. 

- The dialogue for the tour around Hai space should be changed because they repeat almost everything on the planet. "Human Standard" is a cliche in space science fiction and it would be better to avoid it. Tt's never mentioned in the game at all except here. It seems like there has been lots of time for separation between inhabited systems that different languages would be around. There might be a "human standard" language in Endless Sky, but I think it's not a great term. 

- The Syndicate is "famously mysogynistic", and I'm very surprised that a woman became the COO of Syndicated Shipyards. But because she's a big part of the story, it's too much to change and she obviously seems capable of the job. 

- The President of the Republic and the Chancellor of the Free Worlds are two badly introduced parts of the story. These are brand new concepts that isn't ever presented to the player beforehand, but they seem important enough that the player should have known about it already. Presumably I would have heard something about this chancellor while fighting for them! And the two heads of state aren't even named, so they aren't really anything at all. I don't want to know about the politics of the Republic that much.