Change wording and syntax for hai reveal 4 middle
I think that the President of the Republic and the Chancellor of the Free Worlds are two badly made parts of the story. These are brand new concepts and two people that aren't ever presented to the player beforehand, but they seem important enough that the player should have known about it already. Presumably I would have heard something about this Chancellor while fighting for them! Although, I suppose they could have been elected after the war, but I can't imagine it. And the two heads of state aren't even named, so they aren't really anything or anyone at all, they're just there to solve this bit of plot. I don't want to know about the politics of the Republic that much either, so I don't think that there could be that much explanation to be made for them. 

The conversation itself is fine, though.