Fixed Typo in "Remnant: Cognizance 25"
There where two extra words in some of the dialog that didn't make sense, as if the author of the dialog started writing one thing, then changed mid-sentence to write something else. I took out the extra words from the first thing that was written to make the overall sentence make sense.