Add storyline implementation
Adds an implementation of storylines. See issue 7449 (I created this issue, from a different account). The UI is integrated into the log book, and there is also one extra element for visibility which was not in the original issue. Documentation will follow in a later PR, and I will create a plugin to provide an example of a storyline definition.