Let out-of-system escorts without fuel plan their next move
This is a small fix-up of #7379 fixing a new report in #5541.

MoveEscort() has two branches that were blocked if an escort was out of
fuel. These two branches were responsible for planning how an escort
should follow its parent, and already contain code handling the
out-of-fuel condition, by either landing to refuel (via SelectRoute) or
by moving to the system center as a fallback. Also flying through a
wormhole without fuel was blocked as an option as a side effect.

This patch enables taking those branches for escorts without fuel.