Use a PlayerInfo transaction to prevent saving state from mid-conversation
This addresses issue #3746.

The approach taken is pre-emptively saving the PlayerInfo to an
in-memory savegame, and using that state when saving during a
transaction. This should be a robust way of handling this without
adding a lot of infrastructure to keep track of two states.

This patch modifies DataWriter to make it possible to keep a temporary
state.

It still has debugging info.