Short Range Jobs
This adds 4 generic jobs:
- 1 regular delivery
- 1 bulk delivery
- 1 rush delivery
- 1 passenger

All these jobs are restricted to 3 jumps or less distance, and are intended to ensure that the starting player has access to a variety of jobs that are within reach of their starting ship.

No, they don't pay well, but they give the player something that they can actually for sure do instead of starting them off going potential 16 jumps into the unknown. Local traffic is a thing. Not very common, apparently, but still a thing.