BoardingPanel: obscure unknown outfits
when boarding a disabled spaceship, the player might encounter
outfits for the first time.
in that case, it is logical to assume that he doesn't know what they are.
in that case, he will see them labeled as "Unknown <category>" for all categories
beside power where it will be labeled as "Unknown Power Component" until he
reaches an outfitter where he will know exactly what they are.
due to that same fact, the outfit's value is unknown and will be marked in
???

Signed-off-by: Dagg Stompler <daggs@gmx.com>