Add "to accept" ConditionSet
I want to create a repeatable job based on @Hurleveur Unfettered Hai Duel mission framework.

The idea would be to create "Arena Duel" missions that would require the player to have no escorts, and possible be in a certain class of ship.

In order to do this the mission always need to be offered, but not acceptable unless the player meets some preconditions.