Increased Katya's mission chance
OK, so I'm two years into a playthrough and have been intentionally trying to start the FW campaign without success. So I went to take a look at this. This mission needs to spawn early. It is balanced on the expectation that it will. It is written on the expectation that it has. A huge chunk of game content is locked behind it. 

And its spawn chance is 10% base. 

Increased it to 70%, because...why would it be low? Just...why? 

If I have wildly misunderstood how spawn chances work, and they're NOT out of 100, fair enough, but I don't think I have.