BoardingPanel: change logic for unknown outfit
as suggested by tehhowch, for more realistic approach, when trying to
plunder a ship, mark all unknown outfits as such, e.g. even the category
is not known.
in addition, as tehhowch pointed out unknown outfit has direct affect on
the value, so when an unknown outfit is added, use the outfit's mass
as unit value so in the list, the sort will result with cost per mass of
1 and place it at the end of the list.

Signed-off-by: Dagg Stompler <daggs@gmx.com>