Organizing Files
categories, conversations, help, interfaces, ratings, and tooltips are in the UI folder as they all pertain to text presented by the game within the interface and game-controlled elements such as tooltips or warnings.

dialog phrases, effects, governments, persons, and start are in the general folder, as they are actual in-game content that is of a general nature.

intro missions has been moved into the human folder, because it is very exclusively a human mission chain.

This is compatible with but also an alternative to gathering the map related files into a map folder. This doesn't touch any file touched in the map splitting + reorganization PR, but does serve to isolate those files as well.