Heliarch agents have metal headpieces
This is to go with the idea that the Heliarchs' circlets have some manner of neural interface working. The triangular headpieces are for Heliarchs of lesser ranks, but are capable of that just as well. The circlets are reserved for consuls, and have unrestricted access, unlike the headpieces which wouldn't grant one access to every facility.