ci: New ES release process that simplifies the existing one (#8301)
This changes the release process in such a way that it now requires pushing the new version tag manually to the repo to trigger the release workflow that builds all the assets. The workflow will create a draft release, which can be published once the workflow finished uploading all the assets.