Auto-Pilot stops thrusting at ship max speed
Summary
-------

Adds a check to the MoveTo() function so that the auto-pilot is not thrusting if the ship already is at max speed.

* This is adopted from #8308

Code review authorship from the previous PR is preserved.

* Update `source/AI.cpp`.
* Apply suggestions from code review.

Testing Done
------------

It works for most ships.  It can take a while for some ships.

Performance Impact
------------------

N/A

Co-authored-by: Nick <quyykk@protonmail.com>
Co-authored-by: Sam Gleske <875669+samrocketman@users.noreply.github.com>