fix(AI): Fix autopilot for flagship with reverse main thruster (#6604)
This fixes the autopilot logic to move to an object (planet, boarding, ...)
if there is more reverse thrusting than forward thrusting installed.
This includes the special case of having no forward thruster at all.
It also fixes the case of ordering escorts with reverse main thruster
to a specific position (explicitly or implicitly, e.g. while landing).
It does not completely fix #6604, as other problems with escorts having
a reverse main thrusters persist, e.g. gathering. They will be fixed
separately to keep changes small.

Testing of an initial version of this change revealed, that some NPC
ships equipped with a regular thruster and a weaker reverse thruster
relied too much on the reverse thruster. The change depends on a
correct calculation of reverse acceleration, which was wrong (especially
compared with the calculation of regular acceleration).
Therefore, this change fixes the reverse acceleration calculation having
the potential to improve the long-range AI attack behavior for ships
with main reverse thruster it there should be any in the future.