Update human missions.txt
Added Quicksilver Mixup [2] and [3], a continuation for Improvise and bene-dictador's Quicksilver Mixup [0] and [1], named as such for organisational purposes. Their parts are not yet merged, but after they are, i'll update human missions.txt, as such including it between Deal Change 2 and Quicksilver Mixup [2]. Am doing this edit prior to that as to speed the review process of it up, as some showed interest in reviewing the PR now.