Convert NPC Actions from acting on ship events to acting on NPC Triggers
This brings how NPCs handle actions more in line with how Missions handle actions. Currently supported triggers mirror objectives that an NPC block can have and trigger on "completion" of those "objectives." Using quotes, because these don't actually have to be objectives of the NPC in order for them to trigger.