Make new ship names consistent with old ones
The old ship names in "Deep Archaeology 5" had a period after each letter, while the new ones did not have a period after the last letter. I have updated the new ships names to have a period after the last letter so that they are the same as the old ones.