Replace uninhabited spaceport planets with custom ports that recharge all
The purpose of using uninhabited in most of these cases was simply to allow one to refuel but not have services provided. Although the uninhabited keyword will still behave this way when combined with spaceport keyword, might as well make it explicit with the new syntax.

Still left the uninhabited attribute on planets that are truly uninhabited.