Add bounty hunters to Lesath and Shaula
They try to block your path to Greenrock, although they can be bribed... for a steep price.

Also bumped the alternative route payment to 300k.

If you're in a Sparrow, then a single bribe is 132k. Bribe both sets of bounty hunters and you'll make less than if you had just taken the main route. Returns will be worse if you're bribing in a bigger ship.

Reduced the offer chance of the first mission.