Add Mammoth to Advanced Nothern Pirates Sales
the mammoth is a unique moded pirate behemoth that comes up it one signle free world mission. it is not obtainable in any other way.

this pull reqeust adds the mammoth to the sale list of advanced human pirates, since its close to the Betelgeuse Shipyard and it would make sense that these pirates might have captured so and modded them.