allow player to reget bank mortage
If a player doesn't buy a ship and pays off their mortage, the bank won't allow them to get another one, so game over.  This gives the player just enough history for the bank to offer them a loan for the same as the default start again.  And as the history is very rear-loaded, this goes away as soon as the player departs.