No need for player to conceal Coalition...
The player can just talk to Anna about the coalition
without the worry of being overheard. No lore reason why
the Coalition has to be a secret (unlike the remnant),
and anyone passing by close enough to overhear will
dismiss it as something from a novel, game, or work of fiction.