Update "fail" action with no mission name behavior
The documentation for fail is as follows:
> This causes the named mission (or this mission, if no name is given) to fail immediately.

Emphasis on "or this mission, if no name is given". The current behavior is not simply "this mission," but "all missions of this type."

Changed "fail" without a mission name provided to only fail the calling mission. The old behavior of failing all missions of the same type can still be used by explicitly providing the mission name.