Overhaul escort display.
old escort display randomly threw away ships that it didn't want to
display, and left ships uncollapsed if there was room, making it harder
to select them all.

New code implements these merge rules:
* Merge all related types
* Merge unrelated types that are in a different system until there is
  room
* Merge any type until there is room

In addition, the escort display box has been widened slightly and the
escort bars have been shrunk and the padding reduced to make room for
two full columns of escorts, instead of just one. The hitbox for
clicking on the icon has been increased to the full rect, instead of
just the 15 pixel radius circle on the icon.

Fixes #97