Fix catch 22 in new south mission
A week or two ago a series of new missions in the galactic south were committed, however, one of them contained a circular dependency/catch 22 where a mission required itself to be offered in order to offer. It looks like this was supposed to require the previous mission to offer, so that's what it does now.