Fix copyright
Many copyrighted items said that their license was "public domain." This is ambiguous, as they were added to the game after June 2017 and taken from unsplash.com. This means that, depending on when the image was uploaded, it could be either in the public domain or subject to the unsplash license. I have updated the file to add the unsplash license and clarify this. For pictures added to the game before June 2017, they are still listed as public domain with an explanation that they were added before June 2017.