buff scan efficiency
Scan efficiency is now multiplied by 100 before getting the "speed" variable. This essentially means scanning is ten times faster than it was before, however, due to the maximum scan time of ten seconds, this really means that ships can actually have noticeable changes in scan times with more scanners. It is still a "minigame" of trying to stay near the ship, and there is still a noticable difference between ships with a lot of cargo and not much cargo, and the same for outfits. Scan power is untouched. This helps when you need to do that mission to scan a cruiser: you can now do so without it always taking ten seconds. It also makes missions like "corporate espionage" possible.